Update main.js
uniqueDnaTorrance is currently used as global maximum fail count for generating all DNAs. This makes this number dependent on the edition size. It is much more usable if this number is used as maximum try count for a single DNA.

Reset failedCount when creating DNA after each successful DNA creation. Also added try count to log when DNA was created to get a better sense how hard it is to find a valid configuration with more complex filter rules.