Update main.js
I added new layer option allowing to not create any meta data for that layer.

Similar to ignoreAllNamesWithBlank, but for layers.

I did this intentionally with a json setting instead of a file name filter as the former is much cleaner. I would vote for moving all config into the json and not use the filename for this at all.